cmd/trace/exec: Move ExecParser to a common package
